Split feature flag / settings / capabilities seeds out into their own files
